fix: moved getAlertsAndRules inside of useAlerts hook
