+++ packages/ocs/dashboards/persistent-internal/status-card/status-card.tsx
@@ -32,7 +32,7 @@ import {
 } from '@patternfly/react-core';
 import { CephClusterModel } from '../../../models';
 import { DATA_RESILIENCY_QUERY, StorageDashboardQuery } from '../../../queries';
-import { filterCephAlerts, getAlertsAndRules } from '../../../utils/metrics';
+import { filterCephAlerts } from '../../../utils/metrics';
 import { getCephHealthState, getDataResiliencyState } from './utils';
 import { whitelistedHealthChecksRef } from './whitelisted-health-checks';
 import './healthchecks.scss';
@@ -41,8 +41,7 @@ const resiliencyProgressQuery =
   DATA_RESILIENCY_QUERY[StorageDashboardQuery.RESILIENCY_PROGRESS];
 
 export const CephAlerts: React.FC = () => {
-  const [data, loaded, error] = useAlerts();
-  const { alerts } = getAlertsAndRules(data);
+  const [alerts, loaded, error] = useAlerts();
   const filteredAlerts =
     loaded && !error && !_.isEmpty(alerts) ? filterCephAlerts(alerts) : [];
